+++ src/utils/checkUnit.ts
@@ -1,8 +1,0 @@
-import * as vsocde from "vscode";
-import { UnitType } from "../types/index";
-
-const checkUnit = (value: string, unit: UnitType) => {
-  return value.endsWith(unit);
-};
-
-export default checkUnit;